Separates cache operations into their own models
Cache methods are removed from the databse models.

Each endpoint now has its own cache model that handles the getting and
setting of cache data.
