+++ Atelier_postgres_database/models/cacheModels/index.js
@@ -1,0 +1,3 @@
+exports.productsCacheModel = require('./productsCacheModel.js');
+exports.stylesCacheModel = require('./stylesCacheModel.js');
+exports.relatedCacheModel = require('./relatedCacheModel.js');
