--- Atelier_postgres_database/stressTest/loaderio-f860f24adc8e26bc459078499a458e8c.txt
@@ -1,3 +1,0 @@
-loaderio-f860f24adc8e26bc459078499a458e8c
-
-this is the file
